chore: move airbyte type into lib/airbyte
